+++ Model/Context.php
@@ -1,0 +1,47 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\AlekseonEav\Model;
+
+/**
+ * Class Context
+ * @package Alekseon\AlekseonEav\Model
+ */
+class Context extends \Magento\Framework\Model\Context
+{
+    /**
+     * @var \Magento\Store\Model\StoreManagerInterface 
+     */
+    protected $storeManager;
+    
+    /**
+     * Context constructor.
+     * @param \Psr\Log\LoggerInterface $logger
+     * @param \Magento\Framework\Event\ManagerInterface $eventDispatcher
+     * @param \Magento\Framework\App\CacheInterface $cacheManager
+     * @param \Magento\Framework\App\State $appState
+     * @param \Magento\Framework\Model\ActionValidator\RemoveAction $actionValidator
+     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
+     */
+    public function __construct(
+        \Psr\Log\LoggerInterface $logger, 
+        \Magento\Framework\Event\ManagerInterface $eventDispatcher, 
+        \Magento\Framework\App\CacheInterface $cacheManager, 
+        \Magento\Framework\App\State $appState,
+        \Magento\Framework\Model\ActionValidator\RemoveAction $actionValidator,
+        \Magento\Store\Model\StoreManagerInterface $storeManager
+    ) {
+        $this->storeManager = $storeManager;
+        parent::__construct($logger, $eventDispatcher, $cacheManager, $appState, $actionValidator);
+    }
+
+    /**
+     * @return \Magento\Store\Model\StoreManagerInterface
+     */
+    public function getStoreManager()
+    {
+        return $this->storeManager;
+    }
+}